backend/requirements.txt
# --- Latest Flask 3.x Framework Packages ---
Flask~=3.0.3
Flask-SQLAlchemy~=3.1.1
Flask-Migrate~=4.0.7
Flask-Cors~=4.0.1
Flask-Bcrypt~=1.0.1
PyJWT~=2.8.0

# --- Asynchronous Worker for Gunicorn ---
gevent~=24.2.1

# WSGI Server
gunicorn~=22.0.0

# For handling .env files
python-dotenv~=1.0.1